Reversing the direction of handshaking
Since the parent loads the child, and then the child's file loads after, it
makes sense that the child initiates the handshake. Using document.referrer
it sends the handshake message to the parent, which now handshake replies
to the child. This addresses issues where iframe.load was called before
the penpal library was loaded, and so the parent's handshake was premature.